Move user model to correct namespace
